Better synchronization when a worker stops. We make sure that the worker has stopped before restarting the timer. Furthermore, when a worker stops, its last rdb+aof will be renamed to "prev.rdb" and "prev.aof".
This means that, at any given time, there should be two sets of files. The one
which belongs to the current worker (timestamped), and the previous one.

(While the new worker starts up, the old worker continues to process changes
so there should be a few seconds where there's actually 3 sets of files - the
previous worker, the new worker, and the current worker which will shutdown
as soon as the new worker is ready).